Added smoothScatter.2 and tuned plotSignature
The function smoothScatter.2 overrides the default of smoothScatter
which leaves a white frame in the plot around the plotting area. Now the
background is properly drawn to the borders of the plot.

The function plotSignature now takes classes which will be color coded.
Furthermore I adjusted some fontsize and spacing issues. This function
still needs some testing though.